Fixes imports, organizes file structure
